project/requirements.txt
blessings==1.7
bpython==0.17.1
certifi==2018.11.29
chardet==3.0.4
configparser==3.7.3
curtsies==0.3.0
Django==1.11.20
django-bootstrap4==0.0.7
djangorestframework==3.9.1
entrypoints==0.3
enum34==1.1.6
flake8==3.7.6
functools32==3.2.3.post2
greenlet==0.4.15
idna==2.8
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.0
Pygments==2.3.1
pytz==2018.9
requests==2.21.0
six==1.12.0
typing==3.6.6
urllib3==1.24.1
wcwidth==0.1.7
